refactor(docs): apply code-quality review to the landing page
From a 4-angle cleanup review (reuse / simplification / efficiency /
altitude):

- Extract CtaBlock.vue — the hero and closing CTA+install-chip blocks
  were duplicated verbatim and shared one copied ref (one click flashed
  "copied!" on both chips); each instance now owns its state.
- Brackets measure the .cblk spans in the visible code variant instead
  of hardcoded line numbers + a parsed line-height (three reviewers
  flagged this; editing the snippet silently misplaced them before).
- Pre-paint script also sets html.landing-page on /, killing the
  white-docs-page flash on direct loads of the home page.
- injectFonts drops preconnects duplicated from the global head.
- Footer logo references the public /gofish-logo.png instead of a
  byte-identical base64 copy in a .txt fragment.
- Bottle PNG downscaled to display resolution (104KB -> 30KB binary;
  the hero fragment shrinks 598KB -> 487KB) — visually identical.
- DEFAULT_LANG constant in docsLang.ts; sync comments at the two
  sites that cannot import it (inline pre-paint script, CSS :not()
  direction).
- --sheet-w / --card-overlap custom properties tie the gutter-card
  calc()s to the sheet width; --pulltab-grad dedupes the flap gradient.
- Dead CSS removed (.result-card, .tok-com, --persimmon/--plum/--noise
  tokens, subsumed :not(.home) selectors, redundant chip margin,
  duplicated heading typography, dead pageClass frontmatter).

Also: CLAUDE.md now requires screenshots on PRs that add examples or
make big visual docs changes (orphan pr-<n>-assets branch + raw URLs
from the CLI).
