+++ CLAUDE.md
@@ -137,6 +137,7 @@ Key coordinate systems available:
 
 - **Update docs on user-facing API changes**: When you change a public API surface (signatures, option shapes, exported names, default behaviors) in `packages/gofish-graphics/src/`, update the corresponding page under `apps/docs/docs/js/` in the same change — signature, examples, and any equivalences tables. For Python API changes (`packages/gofish-python/`), update the mirror page under `apps/docs/docs/python/`. The docs site has one folder per language (`js/`, `python/`) with a top-level language toggle. Do not defer this to a follow-up.
 - **Update the frontend IR schema when adding an IR-level construct**: Anything that crosses the Python↔JS bridge as serialized IR — a new constraint `type`, mark type, operator, or a new option/field shape — must also be added to the canonical schema in `packages/gofish-ir/src/frontend/`, which has **three encodings that must agree**: the TS type in `schema.ts`, the runtime validator in `validate.ts`, and the emitted JSON Schema in `jsonSchema.ts`. After editing, run `pnpm --filter docs sync-ir-schema` to regenerate the docs page (CI runs `check-ir-schema`). `pnpm --filter @gofish/tests validate-python-ir` validates every Python story's IR against this schema and **will fail CI** if a new construct is missing — so a new constraint type touches the JS factory, the Python wrapper, _and_ these three schema encodings.
+- **Attach screenshots to visual PRs**: Any PR that introduces new examples or makes significant visual changes to the docs site (layout redesigns, new chart renders, styling overhauls) must include screenshots of the result in its description. When working from the CLI (where the GitHub image uploader isn't available), push the PNGs to an orphan `pr-<number>-assets` branch and embed them via `raw.githubusercontent.com` URLs; the branch is never merged and can be deleted after the PR closes.
 - **Monorepo Management**: Uses pnpm workspaces
 - **Visual Development**: Use Storybook (`pnpm storybook`) for interactive development and testing
 - **Iterating on examples with Claude**: `pnpm capture-one "<title/story>"` renders a single
