+++ apps/docs/docs/.vitepress/config.mts
@@ -150,11 +150,15 @@ export default defineConfig({
     // default. On doc pages the route wins; on the home page the saved
     // preference wins. CSS keyed off this attribute does the rest. Internals
     // pages are language-agnostic (the toggle is hidden there), so they fall
-    // through to the saved preference — harmless.
+    // through to the saved preference — harmless. The same script also adds the
+    // `landing-page` class on the home route so the desk/navbar styling paints
+    // immediately on direct loads instead of flashing a stock white docs page.
+    // NOTE: the 'python' fallback literal must match DEFAULT_LANG in
+    // theme/docsLang.ts (this inline script can't import it).
     [
       "script",
       {},
-      `(function(){try{var p=location.pathname,l;if(p.indexOf('/python/')===0){l='python';}else if(p.indexOf('/js/')===0){l='js';}else{var s=localStorage.getItem('gofish-docs-lang');l=(s==='python'||s==='js')?s:'python';}document.documentElement.setAttribute('data-docs-lang',l);}catch(e){}})();`,
+      `(function(){try{var p=location.pathname,l;if(p.indexOf('/python/')===0){l='python';}else if(p.indexOf('/js/')===0){l='js';}else{var s=localStorage.getItem('gofish-docs-lang');l=(s==='python'||s==='js')?s:'python';}document.documentElement.setAttribute('data-docs-lang',l);if(p==='/'||p==='/index.html'){document.documentElement.classList.add('landing-page');}}catch(e){}})();`,
     ],
     ["link", { rel: "preconnect", href: "https://fonts.googleapis.com" }],
     [
