+++ apps/docs/docs/.vitepress/theme/style.css
@@ -199,7 +199,9 @@
 /* Language-aware home hero + toggle. Keyed off <html data-docs-lang>, set
    before first paint by the inline script in config.mts. These rules are
    global on purpose — running them through a component's scoped-CSS transform
-   mangles the nested :not() selector. Python is the default. */
+   mangles the nested :not() selector. The :not([data-docs-lang="js"]) default
+   direction encodes Python as the default — keep in sync with DEFAULT_LANG in
+   docsLang.ts. */
 html[data-docs-lang="js"] .hero-lang-python,
 html:not([data-docs-lang="js"]) .hero-lang-js {
   display: none;
